+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add graph_ops.edge_sub_graph for generating a subgraph of edges.
 
+- Add `return_edge_created_ts` argument to neighbor sampling methods to return timestamps when edges connecting nodes were created.
+
+### Fixed
+- Uniform sampling works in temporal graphs.
+
 ## [0.1.60] - 2022-04-18
 
 ### Added